Basic-bash-script
A bash script that takes an input from the user. After taking the input from the user, the script prints a message using the input it got from the user.
It does this by using the read command. The read command takes an input from the user and stores it in a variable in this script.