Bash-script-case-statement
A bash script with a case statement. This script takes input from a user, and then runs a command depending on the input.
Case statements enables your script to implement a different action depending on the value of a variable or an expression.
For example you may run this script, and depending on the input you give, you will get a different output.